Split math nodes out into MathNodes.kt
